Switch systems/framework from int64_t to int, to better reflect the behavior on 32-bit platforms.  See discussion on #2629.

+++ drake/systems/framework/basic_state_vector.h
@@ -26,7 +26,7 @@ class BasicStateVector : public LeafStateVector<T> {
  public:
   /// Constructs a BasicStateVector that owns a generic BasicVector of the
   /// specified @p size.
-  explicit BasicStateVector(int64_t size)
+  explicit BasicStateVector(int size)
       : BasicStateVector(
             std::unique_ptr<VectorInterface<T>>(new BasicVector<T>(size))) {}
 
@@ -35,9 +35,9 @@ class BasicStateVector : public LeafStateVector<T> {
   explicit BasicStateVector(std::unique_ptr<VectorInterface<T>> vector)
       : vector_(std::move(vector)) {}
 
-  int64_t size() const override { return vector_->get_value().rows(); }
+  int size() const override { return vector_->get_value().rows(); }
 
-  const T GetAtIndex(int64_t index) const override {
+  const T GetAtIndex(int index) const override {
     if (index >= size()) {
       throw std::out_of_range("Index " + std::to_string(index) +
                               " out of bounds for state vector of size " +
@@ -46,7 +46,7 @@ class BasicStateVector : public LeafStateVector<T> {
     return vector_->get_value()[index];
   }
 
-  void SetAtIndex(int64_t index, const T& value) override {
+  void SetAtIndex(int index, const T& value) override {
     if (index >= size()) {
       throw std::out_of_range("Index " + std::to_string(index) +
                               " out of bounds for state vector of size " +
